refactor(repl): extract read_console and history modules from repl/mod.rs
Move read_console_callback, is_r_command_prompt, setup_history,
save_ipc_history and their tests out of the oversized repl/mod.rs
into dedicated sibling modules. The sponge tests move next to
SpongeQueue in state/. Pure code movement, no behavior change.
